+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Vz Keymap extension will be documented in this file.
 
+### [Unreleased]
+- Improved:
+  - The behavior of Undelete (Utrl+U) of a single text inserting into multiple locations with multiple cursors.
+
 ### [0.11.0] - 2020-09-06
 - Added:
   - The undelete stack which allows Ctrl+U to restore previously deleted characters.