+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the Vz Keymap extension will be documented in this file.
 
-### [Unreleased]
+### [0.11.1] - 2020-09-26
 - Improved:
   - The behavior of Undelete (Utrl+U) of a single text inserting into multiple locations with multiple cursors.
   - The behavior of Undelete (Utrl+U) of multiple lines pasting with a single cursor.